chore: renamed poiana package to pigeon.
