+++ pkg/pigeon/service_factory_client.go
@@ -1,4 +1,4 @@
-package poiana
+package pigeon
 
 import "github.com/google/go-github/v50/github"
