+++ pkg/pigeon/service_factory_mock.go
@@ -1,4 +1,4 @@
-package poiana
+package pigeon
 
 import "fmt"
