+++ pkg/pigeon/variables_client.go
@@ -1,4 +1,4 @@
-package poiana
+package pigeon
 
 import (
 	"context"
