+++ pkg/pigeon/variables_mock.go
@@ -1,4 +1,4 @@
-package poiana
+package pigeon
 
 import (
 	"context"
